refactor: move bootstrapping into separate module
Move most of the bootstrapping block in index.ts into its own module,
with a view to further dividing the block into specific modules for
sequelize, nodemailer and logging
